Move custom parser into CLI submodule
