+++ pyproject.toml
@@ -29,6 +29,7 @@ build-backend = "setuptools.build_meta"
 [tool.setuptools]
 packages = [
     "scfw",
+    "scfw.cli",
     "scfw.configure",
     "scfw.loggers",
     "scfw.package_managers",
