+++ scfw/cli/__init__.py
@@ -9,9 +9,9 @@ import sys
 from typing import Callable, Optional
 
 import scfw
+from scfw.cli.parser import ArgumentParser
 from scfw.logger import FirewallAction
 from scfw.package_managers import SUPPORTED_PACKAGE_MANAGERS
-from scfw.parser import ArgumentParser
 
 _LOG_LEVELS = list(
     map(
